+++ core/src/main/java/nl/nn/adapterframework/jms/JMSFacade.java
@@ -16,6 +16,7 @@
 package nl.nn.adapterframework.jms;
 
 import java.io.IOException;
+import java.util.Iterator;
 import java.util.Map;
 
 import javax.jms.ConnectionFactory;
@@ -534,6 +535,9 @@ public class JMSFacade extends JNDIBase implements INamedObject, HasPhysicalDest
 		return send(session, dest, correlationId, message, messageType, timeToLive, deliveryMode, priority, false);
 	}
 	public String send(Session session, Destination dest, String correlationId, String message, String messageType, long timeToLive, int deliveryMode, int priority, boolean ignoreInvalidDestinationException) throws NamingException, JMSException, SenderException {
+		return send(session, dest, correlationId, message, messageType, timeToLive, deliveryMode, priority, ignoreInvalidDestinationException, null);
+	}
+	public String send(Session session, Destination dest, String correlationId, String message, String messageType, long timeToLive, int deliveryMode, int priority, boolean ignoreInvalidDestinationException, Map properties) throws NamingException, JMSException, SenderException {
 		TextMessage msg = createTextMessage(session, correlationId, message);
 		MessageProducer mp;
 		try {
@@ -571,6 +575,14 @@ public class JMSFacade extends JNDIBase implements INamedObject, HasPhysicalDest
 		}
 		if (timeToLive>0) {
 			mp.setTimeToLive(timeToLive);
+		}
+		if (properties!=null) {
+			for (Iterator it = properties.keySet().iterator(); it.hasNext();) {
+				String key = (String)it.next();
+				Object value = properties.get(key);
+				log.debug("setting property ["+name+"] to value ["+value+"]");
+				msg.setObjectProperty(key, value);
+			}
 		}
 		String result = send(mp, msg, ignoreInvalidDestinationException);
 		mp.close();