+++ RELEASES.md
@@ -65,6 +65,7 @@ Upcoming
 - Add support to Larva for Maven based Eclipse projects
 - Sort available scenarios root directories before unavailable ones
 - Adjust the filling of the ESB Common Message Header in the SOAP Header
+- Add copyAEProperties attribute in EsbJmsListener